Fixes pointed out by clang-tidy.

+++ include/depthimage_to_laserscan/DepthImageToLaserScan.hpp
@@ -37,8 +37,8 @@
 #include <cmath>
 #include <string>
 
-#include <depthimage_to_laserscan/depth_traits.hpp>
 #include <depthimage_to_laserscan/DepthImageToLaserScan_export.h>
+#include <depthimage_to_laserscan/depth_traits.hpp>
 #include <image_geometry/pinhole_camera_model.h>
 #include <opencv2/core/core.hpp>
 #include <sensor_msgs/msg/camera_info.hpp>
@@ -181,7 +181,7 @@ namespace depthimage_to_laserscan
       const T* depth_row = reinterpret_cast<const T*>(&depth_msg->data[0]);
       int row_step = depth_msg->step / sizeof(T);
 
-      int offset = static_cast<int>(cam_model.cy() - scan_height / 2);
+      int offset = static_cast<int>(cam_model.cy() - static_cast<double>(scan_height) / 2.0);
       depth_row += offset*row_step; // Offset to center of image
       for(int v = offset; v < offset+scan_height_; v++, depth_row += row_step){
         for(uint32_t u = 0; u < depth_msg->width; u++){ // Loop over each pixel in row
@@ -216,6 +216,6 @@ namespace depthimage_to_laserscan
     int scan_height_; ///< Number of pixel rows to use when producing a laserscan from an area.
     std::string output_frame_id_; ///< Output frame_id for each laserscan.  This is likely NOT the camera's frame_id.
   };
-} // depthimage_to_laserscan
+}  // namespace depthimage_to_laserscan
 
 #endif
